requirements.txt
fastapi
pydantic
bcrypt
pyjwt
python-multipart
email-validator
python-dotenv
requests
sqlalchemy
uvicorn
black
flake8


